document_agent/requirements.txt
a2a-sdk
uvicorn
pydantic
python-dotenv
fastapi
a2a-sdk[http-server]
